+++ example/lib/counter/counter_screen.dart
@@ -9,14 +9,7 @@ class CounterScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return CubeBuilder<CounterCube>(
-      onAction: (cube, action) {
-        if (action is CubeSuccessAction) {
-          print('CubeSuccessAction: ${action.text}');
-        }
-        if (action is CubeErrorAction) {
-          print('CubeErrorAction: ${action.text}');
-        }
-      },
+      onAction: (cube, action) => print(action),
       builder: (context, cube) {
         return Scaffold(
           appBar: AppBar(
@@ -27,9 +20,11 @@ class CounterScreen extends StatelessWidget {
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
                 Text(Cubes.getString('description_counter')),
-                cube.count.build<int>((value) {
-                  return Text(value.toString());
-                }),
+                cube.count.build<int>(
+                  (value) {
+                    return Text(value.toString());
+                  },
+                ),
               ],
             ),
           ),