+++ example/lib/pokemon/pokemon_cube.dart
@@ -3,38 +3,32 @@ import 'package:examplecube/pokemon/repository/model/pokemon.dart';
 import 'package:examplecube/pokemon/repository/pokemon_repository.dart';
 
 class PokemonCube extends Cube {
+  static const LIMIT_PAGE = 20;
   final PokemonRepository repository;
 
   PokemonCube(this.repository);
 
   final list = ObservableList<Pokemon>(value: []);
   final progress = ObservableValue<bool>(value: false);
-  int page = 0;
 
   @override
   void ready() {
-    loadList();
+    fetchPokemonList();
     super.ready();
   }
 
-  void loadList({bool isMore = false}) {
+  void fetchPokemonList({bool isMore = false}) {
     if (progress.value) return;
-    if (isMore) {
-      page++;
-    } else {
-      page = 0;
-    }
-    progress.value = true;
-    repository.getPokemonList(page: page).then((value) {
-      if (isMore) {
-        list.addAll(value);
-      } else {
-        list.value = value;
-      }
-    }).catchError((error) {
-      onAction(CubeErrorAction(text: error.toString()));
-    }).whenComplete(() {
-      progress.value = false;
-    });
+    int page = 0;
+    if (isMore) page = (list.length ~/ LIMIT_PAGE) + 1;
+    progress.update(true);
+    repository
+        .getPokemonList(page: page, limit: LIMIT_PAGE)
+        .then((value) {
+          if (isMore) return list.addAll(value);
+          list.update(value);
+        })
+        .catchError((error) => onAction(CubeErrorAction(text: error.toString())))
+        .whenComplete(() => progress.update(false));
   }
 }