+++ example/lib/pokemon/repository/pokemon_repository.dart
@@ -7,9 +7,8 @@ class PokemonRepository {
   Future<List<Pokemon>> getPokemonList({int page = 0, int limit = 20}) {
     String paramLimit = '';
     if (limit != null) paramLimit = '&limit=$limit';
-    return http.get('http://104.131.18.84/pokemon?page=$page$paramLimit').then(
-        (response) => jsonDecode(response.body)['data']
-            .map<Pokemon>((item) => Pokemon.fromJson(item))
-            .toList());
+    return http.get('http://104.131.18.84/pokemon?page=$page$paramLimit').then((response) {
+      return jsonDecode(response.body)['data'].map<Pokemon>((item) => Pokemon.fromJson(item)).toList();
+    });
   }
 }