+++ lib/src/observable/observable_list.dart
@@ -2,63 +2,75 @@ import 'package:cubes/src/observable/observable_value.dart';
 
 class ObservableList<T> extends ObservableValue<List<T>> {
   ObservableList({Iterable<T> value}) {
-    this.value = value;
+    this.setInitialValue(value);
   }
 
   /// add element in list and notify listeners
   void add(T value) {
     this.value?.add(value);
-    notify();
+    notifyListeners();
   }
 
   /// add element list in list and notify listeners
   void addAll(Iterable<T> value) {
     this.value?.addAll(value);
-    notify();
+    notifyListeners();
   }
 
   /// clear list and notify listeners
   void clear() {
     this.value?.clear();
-    notify();
+    notifyListeners();
   }
 
   /// remove in list and notify listeners
   void remove(T value) {
     this.value?.remove(value);
-    notify();
+    notifyListeners();
   }
 
   /// remove element by index in list and notify listeners
   void removeAt(int index) {
     this.value?.removeAt(index);
-    notify();
+    notifyListeners();
   }
 
   /// remove element in list and notify listeners
   void removeWhere(bool test(T element)) {
     this.value?.removeWhere(test);
-    notify();
+    notifyListeners();
   }
 
   /// remove last element in list and notify listeners
   void removeLast() {
     this.value?.removeLast();
-    notify();
+    notifyListeners();
   }
 
   /// remove element by range in list and notify listeners
   void removeRange(int start, int end) {
     this.value?.removeRange(start, end);
-    notify();
+    notifyListeners();
   }
 
   /// replace elements in list and notify listeners
   void replaceRange(int start, int end, Iterable<T> replacement) {
     this.value?.replaceRange(start, end, replacement);
-    notify();
+    notifyListeners();
   }
+
+  /// get if list is empty
+  bool get isEmpty => this.value?.isEmpty;
+
+  /// get if list is not empty
+  bool get isNotEmpty => !isEmpty;
 
   /// get size list
   int get length => this.value?.length;
+
+  /// get first item
+  T get first => this.value?.first;
+
+  /// get last item
+  T get last => this.value?.last;
 }