+++ lib/src/util/extensions/observable_value_extensions.dart
@@ -1,13 +1,12 @@
 import 'package:cubes/src/observable/observable_value.dart';
-import 'package:cubes/src/observer.dart';
+import 'package:cubes/src/widgets/observer.dart';
 import 'package:flutter/widgets.dart';
 
 extension ObservableValueExtensions on ObservableValue {
   Observer build<T>(
     ObserverBuilder<T> build, {
     bool animate = false,
-    AnimatedSwitcherTransitionBuilder transitionBuilder =
-        AnimatedSwitcher.defaultTransitionBuilder,
+    AnimatedSwitcherTransitionBuilder transitionBuilder = AnimatedSwitcher.defaultTransitionBuilder,
     Duration duration = const Duration(milliseconds: 300),
   }) {
     return Observer<T>(