+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 0.5.1
+
+- Add method `update` and `modify` in `ObservableValue`;
+- [BREAKING CHANGE] not modify directly the `value`, now use `update` or `modify`;
+- Add `isEmpty`, `isNotEmpty`, `first` and `last` in `ObservableList`;
+
 # 0.5.0
 
 - Add CuberProvider. To get the Cube by the children of `CubeBuilder`, `CubeWidget` or `CubeWidgetAnimation` you can use `Cubes.of<MyCube>(context)`;